v1.2

+++ src/task3.cpp
@@ -5,7 +5,7 @@
 
 unsigned long long sumPrime(unsigned int hbound)
 {
-	unsigned int result = 0;
+	unsigned long long result = 0;
 	for (unsigned int i = 2; i < hbound; i++)
 		if (checkPrime(i)) result = result+i;
 	return result;
